Some cleanup of Node and Vue apps.
