+++ example-apps/sre-demo-frontend-app/src/App.vue
@@ -1,23 +1,9 @@
 <script setup>
 import CounterForm from "./components/CounterForm.vue";
-import Version from "./components/Version.vue";
 </script>
 
 <template>
   <main>
-    <div>
-    
-      <div class="container d-flex flex-column">
-        <h3>Front end</h3>
-    VueJS
-
-    <h3>Back end</h3>
-    <Version />
     <CounterForm />
-
-      </div>
-
-    
-    </div>
   </main>
 </template>
